# Conflicts:
#	src/main/java/tj/machines/multi/electric/MetaTileEntityLargeAlloySmelter.java
#	src/main/java/tj/machines/multi/parallel/MetaTileEntityParallelAdvancedLargeChemicalReactor.java
#	src/main/java/tj/machines/multi/parallel/MetaTileEntityParallelElectricBlastFurnace.java
#	src/main/java/tj/machines/multi/parallel/MetaTileEntityParallelLargeArcFurnace.java
#	src/main/java/tj/machines/multi/parallel/MetaTileEntityParallelLargeCentrifuge.java
#	src/main/java/tj/machines/multi/parallel/MetaTileEntityParallelLargeChemicalReactor.java

+++ src/main/java/tj/util/Counter.java
@@ -1,9 +1,9 @@
 package tj.util;
 
 public final class Counter {
-    private int value;
+    private long value;
 
-    public Counter(int initialValue) {
+    public Counter(long initialValue) {
         this.value = initialValue;
     }
 
@@ -15,7 +15,7 @@ public final class Counter {
         this.value -= amount;
     }
 
-    public int getValue() {
+    public long getValue() {
         return this.value;
     }
 }
